requirements.txt
django==1.6.1
django-taggit==0.11.2
south==0.8.4
python-social-auth==0.1.17
psycopg2==2.5.2
PyYAML=3.10